requirements.txt
openai
pyyaml
pygame